Revert "refactor: merge `deno_cache` into `deno_web` (#31183)"
This reverts commit 5eccca8717986ea7643638b41d2a78a818078b61.
