+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "cli/rt",
   "cli/snapshot",
   "ext/bundle",
+  "ext/cache",
   "ext/canvas",
   "ext/cron",
   "ext/crypto",
@@ -86,6 +87,7 @@ denokv_sqlite = { default-features = false, version = "0.12.0" }
 
 # exts
 deno_bundle_runtime = { version = "0.8.0", path = "./ext/bundle" }
+deno_cache = { version = "0.153.0", path = "./ext/cache" }
 deno_canvas = { version = "0.90.0", path = "./ext/canvas" }
 deno_cron = { version = "0.101.0", path = "./ext/cron" }
 deno_crypto = { version = "0.235.0", path = "./ext/crypto" }
