+++ ext/web/benches/encoding.rs
@@ -36,7 +36,6 @@ fn setup() -> Vec<Extension> {
     deno_web::deno_web::init::<Permissions, deno_web::InMemoryBroadcastChannel>(
       Default::default(),
       None,
-      None,
       Default::default(),
     ),
     bench_setup::init(),
