+++ runtime/Cargo.toml
@@ -40,6 +40,7 @@ path = "lib.rs"
 [dependencies]
 deno_ast = { workspace = true, optional = true }
 deno_bundle_runtime.workspace = true
+deno_cache.workspace = true
 deno_canvas.workspace = true
 deno_core.workspace = true
 deno_cron.workspace = true
