+++ runtime/lib.rs
@@ -1,5 +1,6 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
+pub use deno_cache;
 pub use deno_canvas;
 pub use deno_core;
 pub use deno_cron;
