+++ runtime/shared.rs
@@ -10,6 +10,7 @@ extension!(runtime,
     deno_tls,
     deno_web,
     deno_fetch,
+    deno_cache,
     deno_websocket,
     deno_webstorage,
     deno_crypto,
