+++ runtime/worker.rs
@@ -10,6 +10,9 @@ use std::sync::atomic::Ordering;
 use std::time::Duration;
 use std::time::Instant;
 
+use deno_cache::CacheImpl;
+use deno_cache::CreateCache;
+use deno_cache::SqliteBackedCache;
 use deno_core::CompiledWasmModuleStore;
 use deno_core::Extension;
 use deno_core::InspectorSessionKind;
@@ -45,10 +48,7 @@ use deno_process::NpmProcessStateProviderRc;
 use deno_tls::RootCertStoreProvider;
 use deno_tls::TlsKeys;
 use deno_web::BlobStore;
-use deno_web::CreateCache;
 use deno_web::InMemoryBroadcastChannel;
-use deno_web::cache::CacheImpl;
-use deno_web::cache::SqliteBackedCache;
 use log::debug;
 use node_resolver::InNpmPackageChecker;
 use node_resolver::NpmPackageFolderResolver;
@@ -399,12 +399,12 @@ impl MainWorker {
         if elems.len() == 2 {
           let endpoint = elems[0];
           let token = elems[1];
-          use deno_web::cache::CacheShard;
+          use deno_cache::CacheShard;
 
           let shard =
             Rc::new(CacheShard::new(endpoint.to_string(), token.to_string()));
           let create_cache_fn = move || {
-            let x = deno_web::cache::LscBackend::default();
+            let x = deno_cache::LscBackend::default();
             x.set_shard(shard.clone());
 
             Ok(CacheImpl::Lsc(x))
@@ -536,7 +536,6 @@ impl MainWorker {
         >(
           services.blob_store.clone(),
           options.bootstrap.location.clone(),
-          create_cache,
           services.broadcast_channel.clone(),
         ),
         deno_fetch::deno_fetch::args::<PermissionsContainer>(
@@ -551,6 +550,7 @@ impl MainWorker {
             ..Default::default()
           },
         ),
+        deno_cache::deno_cache::args(create_cache),
         deno_websocket::deno_websocket::args::<PermissionsContainer>(),
         deno_webstorage::deno_webstorage::args(
           options.origin_storage_dir.clone(),
@@ -1058,6 +1058,7 @@ fn common_extensions<
     deno_webgpu::deno_webgpu::init(),
     deno_canvas::deno_canvas::init(),
     deno_fetch::deno_fetch::lazy_init::<PermissionsContainer>(),
+    deno_cache::deno_cache::lazy_init(),
     deno_websocket::deno_websocket::lazy_init::<PermissionsContainer>(),
     deno_webstorage::deno_webstorage::lazy_init(),
     deno_crypto::deno_crypto::lazy_init(),
